Added JavaScript functionality
Setup local storage called "tasks". This is where all tasks will be stored locally

Setup an event listener that prevents the form from refreshing on submit

Created a const task object literal that targets the task input field as well as the category radio button

Set the task object literal to push to the tasks array. (Did this with JSON.stringify to store tasks as in a JSON format)

Used e.target.reset() to clear the input and radio field after the submit button is clicked. Thiw will result in in the input fields being empty after just adding a task